requirements.txt
google-genai
pdf2image
Pillow
python-dotenv
pypdf
